Code cleanup
Got rid of the images_s struct stuff for drawing bitmaps, back to the
normal way of calling draw_bitmap(...). It was an experiment to see if
it produced faster/smaller code when passing a pointer to a struct with
all the data instead of a bunch of func args. Ended up with the code
being slightly slower and larger.